add ae2, ae2-cc compat, ae2 wireless, gravestone
also adding config for gravestone mod so items drop on the ground instead
of going into your inventory when you break the gravestone.

Adding charm config as well, as it was previously modified on the server.
this way it's easier for other people to see and make
changes if needed.